feat: :sparkles: add satellite info boxes to mobile settings manager
